+++ src/civilization/general/components.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::{Component, Entity, Reflect};
 use bevy::utils::HashMap;
 
-#[derive(Component, Debug)]
+#[derive(Component, Debug, Reflect)]
 pub struct Area {
     pub max_population: u8,
 }
@@ -28,4 +28,19 @@ pub struct StartArea;
 #[derive(Component, Debug, Reflect)]
 pub struct Token {
     pub player: Entity,
+}
+
+#[derive(Component, Debug, Reflect)]
+pub struct Stock {
+    pub max_tokens: usize,
+    pub tokens: Vec<Entity>,
+}
+
+impl Stock {
+    pub(crate) fn new(max_tokens: usize, tokens: Vec<Entity>) -> Self {
+        Stock {
+            max_tokens,
+            tokens,
+        }
+    }
 }
