+++ src/civilization/general/mod.rs
@@ -1,3 +1,4 @@
 pub mod plugin;
 pub mod systems;
 pub mod components;
+pub mod events;
