+++ src/civilization/population_expansion/events.rs
@@ -1,0 +1,4 @@
+use bevy::prelude::Event;
+
+#[derive(Event, Debug)]
+pub struct StartManualPopulationExpansionEvent;
